% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{addNewData}
\alias{addNewData}
\title{Modifies 'data' by adding new values supplied in newDataFileName}
\usage{
addNewData(data, newDataFileName)
}
\arguments{
\item{data}{existing data.frame}

\item{newDataFileName}{name of lookup table}
}
\value{
modified data.frame
}
\description{
newDataFileName is expected to have columns 
c(lookupVariable,lookupValue,newVariable,newValue,source)
}
\details{
Within the column 'newVariable', replace values that
match 'lookupValue' within column 'lookupVariable' with the value
newValue'.  If 'lookupVariable' is NA, then replace *all* elements
of 'newVariable' with the value 'newValue'.

Note that lookupVariable can be the same as newVariable.
}

