% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{readNewData}
\alias{readNewData}
\title{Utility function to read/process newDataFileName for addNewData}
\usage{
readNewData(newDataFileName)
}
\arguments{
\item{newDataFileName}{name of lookup table}
}
\value{
data.frame with columns c(lookupVariable,lookupValue,newVariable,newValue,source)
}
\description{
Utility function to read/process newDataFileName for addNewData
}

